requirements.txt
streamlit==1.31.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
matplotlib==3.8.3
torch==2.2.2
joblib==1.3.2
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
sqlalchemy==2.0.15
pydantic==2.0.0
pydantic-settings==2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python==4.7.0
numpy==1.24.3
python-dotenv==1.0.0
requests==2.28.2
jinja2==3.1.2